Simpler and more Rubyish API
Calling like `"Chris".possessive("AP", "Chris")` is a little repetitive, since you have to specify "Chris", so we'll assume that `.possession` calls operate on the string they are called on.

And, because we are lazy, let's use symbols instead of strings to specify the possessive style. It's also more idiomatic ruby, I think.